transfer types.remove lab.mcp
