--- src/lab/mcp/__init__.py
@@ -1,36 +1,0 @@
-from __future__ import annotations
-
-from ._typing import (
-    AudioPart,
-    ContentPart,
-    ConversationState,
-    FilePart,
-    ImagePart,
-    ImageRefResult,
-    ImageURL,
-    OpenAIContent,
-    OpenAIMessage,
-    ScreenShotResult,
-    TextPart,
-    ToolCallLike,
-    ToolTraceItem,
-)
-from .util import dump_openai_msg, prompt_result_to_text
-
-__all__ = [
-    "ToolTraceItem",
-    "ToolCallLike",
-    "ScreenShotResult",
-    "ConversationState",
-    "dump_openai_msg",
-    "prompt_result_to_text",
-    "OpenAIMessage",
-    "ImageRefResult",
-    "TextPart",
-    "ImagePart",
-    "AudioPart",
-    "FilePart",
-    "ContentPart",
-    "OpenAIContent",
-    "ImageURL",
-]
